[Xamarin.Android.Build.Tasks] support $(AndroidUseLatestPlatformSdk) in binding projects
After API 31 is stable, binding projects with
`TargetFrameworkVersion=11.0` and `AndroidUseLatestPlatformSdk=true`
fail to build with:

    Task ResolveNuGetPackageAssets
    ...
    Microsoft.NuGet.targets(198,5): Your project does not reference "MonoAndroid,Version=v12.0" framework. Add a reference to "MonoAndroid,Version=v12.0" in the "TargetFrameworks" property of your project file and then re-run NuGet restore.

What happened is that `$(TargetFrameworkVersion)` was changed to 12.0
during `Build`, but not during `Restore`. So NuGet restored with
`MonoAndroid,Version=v11.0` instead of `MonoAndroid,Version=v12.0`.

I found that we have this setup for other project types, but not
binding projects:

    <ResolveNuGetPackageAssetsDependsOn>
      _SetLatestTargetFrameworkVersion;
      $(ResolveNuGetPackageAssetsDependsOn)
    </ResolveNuGetPackageAssetsDependsOn>

And so the `_SetLatestTargetFrameworkVersion` was not running at the
appropriate time for binding projects.